Separate collection, card, and practice_session tests
